requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
dj-database-url==2.0.0
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.9
